fix: correct schema and model issues for team chore tables
- Update base CREATE TABLE allowance_chores to include team columns
- Remove redundant index on completion_id (covered by UNIQUE constraint)
- Change MinTeamSize from int to int64 for consistency with other ID fields
- Add allowance_team_completions table to test DB schema

Bead: Hytte-jzbr

+++ internal/allowance/models.go
@@ -15,7 +15,7 @@ type Chore struct {
 	Active           bool    `json:"active"`
 	CreatedAt        string  `json:"created_at"`
 	CompletionMode   string  `json:"completion_mode"`  // solo, team
-	MinTeamSize      int     `json:"min_team_size"`    // minimum participants for team mode
+	MinTeamSize      int64   `json:"min_team_size"`    // minimum participants for team mode
 	TeamBonusPct     float64 `json:"team_bonus_pct"`   // bonus percentage awarded for team completion
 }
